+++ jedi_language_server/jedi_utils.py
@@ -190,6 +190,7 @@ def complete_sort_name(name: Completion) -> str:
     Should be passed to the sortText field in CompletionItem. Strings sort a-z,
     a comes first and z comes last.
     """
-    if name.type == "param":
+    if name.type == "param" and name.name.endswith("="):
+        # params directly associated with filling out current function sig
         return f"a{name.name}"
     return f"z{name.name}"